+++ lib/helpers/accumulator.dart
@@ -1,7 +1,20 @@
-class AccumulatorList<T, V> {
-  final Map<T, Set<V>> values = {};
+import 'package:money/models/date_range.dart';
 
-  void cumulate(T key, V value) {
+/// A generic class that accumulates values of type `V` associated with keys of type `T`.
+/// It uses a `Map<K, Set<V>>` to store the accumulated values.
+class AccumulatorList<K, V> {
+  final Map<K, Set<V>> values = {};
+
+  void clear() {
+    values.clear();
+  }
+
+  /// Adds a value of type `V` to the set associated with the provided `key` of type `K`.
+  ///
+  /// If the `key` already exists in the `values` map, it retrieves the existing set
+  /// and adds the `value` to that set. If the `key` doesn't exist, it creates a new
+  /// set containing the `value` and associates it with the `key` in the `values` map.
+  void cumulate(K key, V value) {
     if (values.containsKey(key)) {
       final existingSet = values[key] as Set<V>;
       existingSet.add(value);
@@ -11,15 +24,25 @@ class AccumulatorList<T, V> {
     }
   }
 
-  List<T> getKeys() {
+  bool containsKey(K key) {
+    return values.containsKey(key);
+  }
+
+  /// Returns a list of all the keys present in the `values` map.
+  List<K> getKeys() {
     return values.keys.toList();
   }
 
-  List<V> getList(T key) {
+  /// Retrieves the set of values associated with the provided `key`.
+  ///
+  /// If the `key` exists in the `values` map, it converts the set to a list and returns it.
+  /// If the `key` doesn't exist, it returns an empty list.
+  List<V> getList(K key) {
     return values[key]?.toList() ?? [];
   }
 }
 
+/// Tally values
 class AccumulatorSum<K, V> {
   final Map<K, V> values = {};
 
@@ -35,6 +58,10 @@ class AccumulatorSum<K, V> {
     } else {
       values[key] = value;
     }
+  }
+
+  bool containsKey(K key) {
+    return values.containsKey(key);
   }
 
   K? getKeyWithLargestSum() {
@@ -68,5 +95,51 @@ class AccumulatorSum<K, V> {
 
   List<MapEntry<K, V>> getEntries() {
     return values.entries.toList();
+  }
+}
+
+/// Tally values
+class AccumulatorDateRange<K> {
+  final Map<K, DateRange> values = {};
+
+  void clear() {
+    values.clear();
+  }
+
+  void cumulate(K key, DateTime value) {
+    if (values.containsKey(key)) {
+      values[key]!.inflate(value);
+    } else {
+      values[key] = DateRange()..inflate(value);
+    }
+  }
+
+  bool containsKey(K key) {
+    return values.containsKey(key);
+  }
+
+  DateRange? getValue(K key) {
+    return values[key];
+  }
+
+  List<MapEntry<K, DateRange>> getEntries() {
+    return values.entries.toList();
+  }
+}
+
+/// Ensure unique Key [K] instances, that cumulate unique instance of [I] another accumulator
+/// this last accumulator stores [V]
+class MapAccumulator<K, I, V> {
+  Map<K, AccumulatorSum<I, V>> map = {};
+
+  void cumulate(K k, I i, V v) {
+    if (!map.containsKey(k)) {
+      map[k] = AccumulatorSum<I, V>();
+    }
+    map[k]!.cumulate(i, v);
+  }
+
+  AccumulatorSum<I, V>? getLevel1(K key) {
+    return map[key];
   }
 }