+++ lib/views/adaptive_view/adaptive_list/transactions/list_view_transactions.dart
@@ -4,8 +4,8 @@ import 'package:money/models/fields/field_filter.dart';
 import 'package:money/models/money_objects/money_objects.dart';
 import 'package:money/models/money_objects/transactions/transaction.dart';
 import 'package:money/storage/data/data.dart';
+import 'package:money/views/adaptive_view/adaptive_list/adaptive_columns_or_rows_single_seletion.dart';
 import 'package:money/views/view_transactions/dialog_mutate_transaction.dart';
-import 'package:money/views/adaptive_view/adaptive_list/list_view.dart';
 
 class ListViewTransactions extends StatefulWidget {
   final List<Field> columnsToInclude;
@@ -35,11 +35,6 @@ class _ListViewTransactionsState extends State<ListViewTransactions> {
   late int selectedItemIndex = widget.selectedItemIndex;
 
   @override
-  void initState() {
-    super.initState();
-  }
-
-  @override
   Widget build(final BuildContext context) {
     // get the list sorted
     final List<Transaction> transactions = widget.getList();
@@ -49,52 +44,37 @@ class _ListViewTransactionsState extends State<ListViewTransactions> {
 
     MoneyObjects.sortList(transactions, widget.columnsToInclude, sortBy, sortAscending);
 
-    return Column(
-      children: <Widget>[
-        // Table Header
-        MyListItemHeader<Transaction>(
-          columns: widget.columnsToInclude,
-          filterOn: FieldFilters(),
-          sortByColumn: sortBy,
-          sortAscending: sortAscending,
-          onTap: (final int index) {
-            setState(() {
-              if (sortBy == index) {
-                // same column tap/click again, change the sort order
-                sortAscending = !sortAscending;
-              } else {
-                sortBy = index;
-              }
-              widget.onUserChoiceChanged?.call(sortBy, sortAscending, selectedItemIndex);
-            });
-          },
-        ),
-        // Table list of rows
-        Expanded(
-          child: MyListView<Transaction>(
-            fields: Fields<Transaction>()..setDefinitions(widget.columnsToInclude),
-            list: transactions,
-            displayAsColumn: true,
-            selectedItemIds: ValueNotifier<List<int>>([selectedItemIndex]),
-            onSelectionChanged: (final int uniqueId) {
-              setState(() {
-                selectedItemIndex = uniqueId;
-                widget.onUserChoiceChanged?.call(sortBy, sortAscending, uniqueId);
-              });
-            },
-            onLongPress: (final BuildContext context2, final int uniqueId) {
-              final Transaction instance = findObjectById(uniqueId, transactions) as Transaction;
-              showTransactionAndActions(
-                context: context2,
-                transaction: instance,
-              ).then((value) {
-                selectedItemIndex = uniqueId;
-                widget.onUserChoiceChanged?.call(sortBy, sortAscending, selectedItemIndex);
-              });
-            },
-          ),
-        ),
-      ],
+    return AdaptiveListColumnsOrRowsSingleSelection(
+      list: transactions,
+      fieldDefinitions: widget.columnsToInclude,
+      filters: FieldFilters(),
+      sortByFieldIndex: sortBy,
+      sortAscending: sortAscending,
+      selectedId: -1,
+      // Field & Columns
+      displayAsColumns: true,
+      backgoundColorForHeaderFooter: Colors.transparent,
+      onColumnHeaderTap: (final int index) {
+        setState(() {
+          if (sortBy == index) {
+            // same column tap/click again, change the sort order
+            sortAscending = !sortAscending;
+          } else {
+            sortBy = index;
+          }
+          widget.onUserChoiceChanged?.call(sortBy, sortAscending, selectedItemIndex);
+        });
+      },
+      onItemLongPress: (final BuildContext context2, final int uniqueId) {
+        final Transaction instance = findObjectById(uniqueId, transactions) as Transaction;
+        showTransactionAndActions(
+          context: context2,
+          transaction: instance,
+        ).then((value) {
+          selectedItemIndex = uniqueId;
+          widget.onUserChoiceChanged?.call(sortBy, sortAscending, selectedItemIndex);
+        });
+      },
     );
   }
 }